+++ src/auth/auth.module.ts
@@ -5,12 +5,15 @@ import { FormioModule } from '../formio.module';
 import { FormioAuthComponent } from './auth.component';
 import { FormioAuthLoginComponent } from './login/login.component';
 import { FormioAuthRegisterComponent } from './register/register.component';
+import { FormioAuthRouteConfig } from './auth.config';
+import { FormioAuthRoutes } from './auth.routes';
+import { extendRouter } from '../formio.utils';
 
 @NgModule({
   imports: [
     CommonModule,
     FormioModule,
-    RouterModule.forChild([])
+    RouterModule
   ],
   declarations: [
     FormioAuthComponent,
@@ -18,4 +21,11 @@ import { FormioAuthRegisterComponent } from './register/register.component';
     FormioAuthRegisterComponent
   ]
 })
-export class FormioAuth {}
+export class FormioAuth {
+  static forRoot(config?: FormioAuthRouteConfig): any {
+    return extendRouter(FormioAuth, config, FormioAuthRoutes);
+  }
+  static forChild(config?: FormioAuthRouteConfig): any {
+    return extendRouter(FormioAuth, config, FormioAuthRoutes);
+  }
+}