+++ src/manager/index.ts
@@ -1,0 +1,10 @@
+export { FormManagerConfig } from './form-manager.config';
+export { FormManagerService } from './form-manager.service';
+export { FormManagerRoutes } from './form-manager.routes';
+export { FormManagerIndexComponent } from './index/index.component';
+export { FormManagerFormComponent } from './form/form.component';
+export { FormManagerViewComponent } from './view/view.component';
+export { FormManagerEditComponent } from './edit/edit.component';
+export { FormManagerDeleteComponent } from './delete/delete.component';
+export { FormManagerModule } from './form-manager.module';
+