+++ webpack.config.ts
@@ -77,7 +77,7 @@ export default (environment = 'development') => {
                 ENV: JSON.stringify(environment)
             }),
             new webpack.ContextReplacementPlugin(
-                /angular(\\|\/)core(\\|\/)@angular/,
+                /\@angular(\\|\/)core(\\|\/)fesm5/,
                 path.join(__dirname, 'src')
             ),
             new HtmlWebpackPlugin({